[Blink] Make GetOrCreateResourceProvider() Canvas2D/ImageBitmap-only
This CL changes OffscreenCanvas::GetOrCreateResourceProvider() to
OffscreenCanvas::GetOrCreateResourceProviderForCanvas2DOrImageBitmap().
As the rename makes clear, the method can be called only for those two
context types.

The CL also simplifies the method for the fact that it cannot be called
for WebGL or WebGPU.

Bug: 352263194
Change-Id: I6f60bf42f1466795189005a4f186574c605d5738
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625059
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1471621}
NOKEYCHECK=True
GitOrigin-RevId: e52fdc51db3cf3f997544b3c715a699de475fa3d