Enable DisplayAudioCaptureKillSwitch by default.
This effectively activates the kill switch that was introduced in
crrev.com/c/6450578. The change caused unwanted side effects: echo
cancellation was removed for display audio sources. Although this is an
unintended feature and not according to spec, it's been used by some
applications, which is why the change is reverted for now.

Bug: 410466097, 422611724
Change-Id: I5c46dd24a488d9e95ba802bf2fe6f4a2a2f8d7f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625664
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1471698}
NOKEYCHECK=True
GitOrigin-RevId: 08d3d122d97af816f5f40749e6bb8b9533d4d3ca