[Masonry] Row grid axis item contribution size
This change updates all column-based tests to start with column-* and
adds row test counterparts, all named to start with row-*.

In writing tests, there were two classes of track sizing bugs found so
far. One was that we were using the min/max size contributions toward
track sizing in both the column and row axis. This change adds similar
code to Grid item contribution calculations by adding a way to get the
block size contribution for virtual items when sizing a row grid axis.

To accomplish this, we now cache both the min/max contribution sizes and
block contribution size on the GridItem for virtual masonry items when
appropriate. To calculate the block contribution, this change adds a
ComputeMasonryItemBlockContribution() method, that calls into a new
LayoutMasonryItemForMeasure(), which is similar to how Grid handled this
case, as well.

This code doesn't yet handle margins, baselines or submasonry with
various TODOs added throughout.

The second class of bug found so far is exercised in
wpt_internal/css/css-masonry/row-explicit-placement-001.html. We
currently are sizing the rows too large in this case. This is added to
TestExpectations as a failure and will be fixed in a follow up change.

Bug: 343257585
Change-Id: If9d7feabc866c9ed9da3f62780052c701f17c293
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616763
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472149}
NOKEYCHECK=True
GitOrigin-RevId: 7bdbdc0ca97f58f4cb45eaf6424e105a86f0bf67