Local Storage: Remove partition.mojom
This change removes the partition.mojom between the browser and storage
service. In 2019, it was envisioned that multiple clients could be bound
to a single PartitionImpl. However, this is not the case today. The only
client in production code is the storage_partition_impl class in the
content layer which only ever has a single member of type:
`mojo::Remote<storage::mojom::Partition> partition_;`.

This support for multiple clients is obscuring the true source of
reliability issues in LocalStorage without actually being used in
production code. So, this change removes the partition.mojom and has the
storage_service.mojom directly manage binding Local and Session Storage.

Bug: 396030877
Change-Id: I081c6b4a8d7f5f88a8c1c5d4f06e31ce0d339c2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6572332
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Evan Stade <evanstade@microsoft.com>
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1472219}
NOKEYCHECK=True
GitOrigin-RevId: ae01f8bcc0026b03297501e7e5121056be566044