Add a fast-reject filter for the subject of selectors.
In a complex selector like .a .b .c .d#e, we would have bucketing
to quickly reject (not even look at) every element that didn't
match #e, and a Bloom filter to reject elements that don't have
.a, .b and .c as ancestors, but .d would need to go through the
regular selector matching (possibly easy selector matching).

However, we now have a Bloom filter on every element, containing
bits for the hashes for all classes and attributes set on that
element. We can use that for a fast-reject filter similar to the
ancestor Bloom filter; we store a bitmask on RuleData. For sites
that use multiple such selectors in the subject, this is a nice
win, and on others, it doesn't cost much, since the test is
extremely cheap (e.g., on x86, it reduces to two instructions:
TEST [mem],reg + JZ target). The downside, of course, is that we
grow RuleData by 32 bits for the filter.

Style perftest (Zen 3, LTO but no PGO):

  Initial style (µs)     Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                5455      5389   +1.2%  [ +0.9%,  +1.5%]
  Encyclopedia            42259     41730   +1.3%  [ +0.3%,  +2.2%]
  Extension               55573     54875   +1.3%  [ +1.0%,  +1.6%]
  News                    21949     21808   +0.6%  [ +0.3%,  +0.9%]
  Search                   7690      7624   +0.9%  [ +0.5%,  +1.2%]
  Social1                 13661     13647   +0.1%  [ -0.2%,  +0.4%]
  Social2                  8652      8562   +1.1%  [ +0.7%,  +1.7%]
  Sports                  26752     26461   +1.1%  [ +0.2%,  +2.0%]
  Video                   17724     17532   +1.1%  [ +0.7%,  +1.5%]
  Geometric mean                            +1.0%  [ +0.7%,  +1.2%]

  Recalc style (µs)      Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                2923      2925   -0.1%  [ -0.4%,  +0.3%]
  Encyclopedia            33593     33510   +0.2%  [ -0.1%,  +0.6%]
  Extension               43911     43604   +0.7%  [ +0.3%,  +1.1%]
  News                    10965     10858   +1.0%  [ +0.6%,  +1.3%]
  Search                   2822      2792   +1.1%  [ +0.6%,  +1.5%]
  Social1                  6795      6753   +0.6%  [ +0.3%,  +1.0%]
  Social2                  4182      4164   +0.4%  [ +0.0%,  +0.8%]
  Sports                  12639     12572   +0.5%  [ +0.2%,  +0.9%]
  Video                    7386      7054   +4.7%  [ +4.3%,  +5.1%]
  Geometric mean                            +1.0%  [ +0.8%,  +1.3%]

The memory profile is a bit confusing. Sports and Video are listed
as using significantly more GCAlloc, but even Sports only has ~25k
RuleData entries in all, so the growth should be ~100 kB, not 431 kB.
We will need to keep a watch on this.

  ECommerce       2402 →  2401 kB  (  -1 kB)
  Encyclopedia:  10354 → 10376 kB  ( +22 kB)
  Extension      11380 → 11383 kB  (  +3 kB)
  News:           6642 →  6649 kB  (  +7 kB)
  Search:         3686 →  3689 kB  (  +3 kB)
  Social1:        6067 →  6099 kB  (  +2 kB)
  Social2:        3282 →  3271 kB  ( -11 kB)
  Sports:        13130 → 13561 kB  (+431 kB)
  Video:         12084 → 12293 kB  (+209 kB)

Speedometer3 is neutral, presumably since it doesn't use too many
such selectors.

Change-Id: Icc95ebf16a082f857c2d6494db4b909665c5fded
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621029
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472268}
NOKEYCHECK=True
GitOrigin-RevId: 97586ea8d4e5239ede40a6c6e5030fe99fa6fd60