[PEPC] Set permission element's baseline to be in the middle
The permission element is not centered relative to a pre existing
content, when the icon is added to the button. This is a baseline issue.
The baseline shifts when the icon is added. Set vertical-align to middle
to always have the button centered.

Bug: 423840820
Change-Id: I9339edc04c0b68c51087bca001d8e576fa9851a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6633329
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Ravjit Uppal <ravjit@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472317}
NOKEYCHECK=True
GitOrigin-RevId: e0b312fa4ac775bfaada14206e512e6519cd8325