[gap-decorations] Marking row-rule-color as interpolable: false
This CL marks row-rule-color as not interpolable, since we have not
implemented this functionality for it yet. This is consistent with
row-rule-width.

This CL also makes sure we don't allow `SetVisitedColor` for now
for kRowRuleColor, similar to kAccentColor.

Neither of these changes change existing behavior, but we are
confident these will fix a crasher we have seen in Microsoft Edge,
for which we have not been able to find a repro case.

The added test hits a crash that is very similar to the crash stacks
we observed, but not the same.

Bug: 357648037
Change-Id: I2baf2b33a28b41915db544336533c033eb123ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6635191
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1472491}
NOKEYCHECK=True
GitOrigin-RevId: 10b051a6d612ad2957c30c1574ec8761bd98232e