Add `DebugName` to LargestTextCandidate trace events
In a previous CL (crrev.com/c/6491448) we added this output to a bunch
of trace events that have an associated node, but that did not include
largest text candidates. This name is useful as we can show it on the
frontend to make it more clear which node we are referring to -
especially if the trace was imported so we can't use CDP To look up the
right node ID.

If required, I will send another CL to make a similar change for
LargestPaintCandidate events.

Bug: 415952073
Change-Id: I71bc50e2cab45eec3cc08c98d9b12ae661221e47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625582
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472500}
NOKEYCHECK=True
GitOrigin-RevId: 4959d9320f1b7220f51cc6eabb567e8449ebe045