[gap-decorations] Restrict width animation to just single values
This CL makes sure that we don't try to access GetLegacyValue for
ColumnRuleWidth and RowRuleWidth when they have more than a single
value.

This is because currently, there are test cases that might lead us down
this route and hit a CHECK when we try to query the legacy value of a
repeater.

This doesn't regress existing behavior, since this only applies for
repeater values for ColumnRuleWidth, which is only applicable with
`CSSGapDecorations`. Today we only support animation for single values
and not lists or repeaters.

Bug: 357648037
Fixed: 422814958
Change-Id: I98ce58d43e03836e209365771231bfd3fdc049d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625244
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1472684}
NOKEYCHECK=True
GitOrigin-RevId: 97c82f75ec4d6d4e2626727653ae660fc4a45ec5