[CodeHealth] Modernising script downloader to use optional
`ModuleScriptDownloader` is using a deprecated variant of
`SimpleUrlLoader` that rely on a `unique_ptr` to communicate the
presence of a response body. This change modernises the interface of
this type to rely on a `std::optional<std::string>` to transmit the
response body.

Bug: 410044687
Change-Id: I08af89e5da77927dfe90a77f338a9d5add21a02d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6629630
Reviewed-by: Cammie Smith Barnes <cammie@chromium.org>
Commit-Queue: Claudio DeSouza <cdesouza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1472783}
NOKEYCHECK=True
GitOrigin-RevId: 0e75a5186bbaea03f18c80d1d2fe846050232b5a