corner-shape: improve outline rendering, including dotted
Instead of relying on a rounded stroke, which appears both
inconsistent with square borders and also doesn't work for
dotted/dashed, build a path specific for the case where the rect is
outside of the origin rect, by continuing the curves
with a straight line until they reach the outer bound.

To continue the curve with a straight line, we compute the quadratic
control point of a similar curve, and draw a straight line from there.
This would be close to the tangent of the curve.

The resulting outlines are a bit more straight, but also
more consistent.

The path clipping for color joins needed to expand to include the
outline.

Bug: 417420173
Change-Id: I7d51979e0c2aac789f13c9a7d7361a071e53e22a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6624859
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@google.com>
Cr-Commit-Position: refs/heads/main@{#1472956}
NOKEYCHECK=True
GitOrigin-RevId: dc42b4d6a4d9727e07b23245b6b8b3369dc0ef51