[scoped-VT] Add test with negative playback-rate
Ensure that the duration of the transition is effectively shortened by
setting a negative playback-rate. Unlike reverse, which will seek to
the end time for the animation, setting the playback rate doesn't
change current time and the animation ends with the next update.

Bug: 420690560
Change-Id: I7a584f6e26d98523e2f2b100ac301cdaa5050251
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6637898
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473165}
NOKEYCHECK=True
GitOrigin-RevId: ca9d80c64ed870b3eeafae67d8fe5913c290bcd9