HasAnchorEvaluator should not be considered for diffing
HasAnchorEvaluator flag was part of style diffing, which meant there was
always a diff between the standard recalc and the interleaved one. In
particular that made a difference for highlight styles which triggers a
recalc for children in the interleaved style recalc.

The interleaved recalc for anchored element descendants during layout is
a more general problem. Currently, it affects transitions of explicitly
inherited properties which are allowed to change for position-try
fallbacks, but with anchored() container queries it will become a more
general issue which needs to be addressed for ShouldStoreOldStyle() in
StyleResolver.

Make the ComputedStyle field custom_compare to avoid diffs for
HasAnchorEvaluator.

Bug: 409220513
Change-Id: I6581ec6097f4fabcafcd63606558f3cca8a886aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6641261
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473399}
NOKEYCHECK=True
GitOrigin-RevId: 54a1dcb2d2f484d119b921271b539da11ecea405