[SVG] Granular handling of mutations when attributes are changed on referenced svg element.
According to the spec [1], any event listeners defined on an
element in the referenced graphics must also listen for the same
event, at the same capture phase, on each instance of that element
in a use-element shadow tree. In Chromium, we handle this by copying
the event listeners of the target element to the corresponding
shadow DOM element.

However, if multiple events are defined and we mutate the referenced
element using those events, an issue arises where subsequent event
states get lost. The reason for this is the way we currently handle
mutations in Chromium. We handle them by marking all the related
use-instances as dirty (and eventually deleting them), which means
all state relevant to specific nodes gets lost.

To remedy this in this CL, we introduced a more fine-grained
handling of mutations instead of rebuilding the use-shadow DOM
tree from scratch. Here, when attributes on an referenced svg
element are changed or modified, we pass or supply the same change
to all the shadow instances of the element as well.
This way, the original problem of events getting lost will also be
fixed as we are not rebuilding the use-shadow tree.

[1] https://svgwg.org/svg2-draft/struct.html#UseEventHandling

Bug: 416091474
Change-Id: I33c1133faf0c6664b2500b67383bd2c2a8c61e58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6623901
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1473451}
NOKEYCHECK=True
GitOrigin-RevId: 1c060c16e696c82bd2d2b01886a67a63423d90c1