[menu]: Implement checkability and fieldset integration
This CL introduces the `checkable` content attribute to the `<fieldset>`
element, with the values `single` and `multiple`, which control
menuitem exclusivity.

This CL un-[Reflect]s the menu item's `checked` IDL attribute, so that
this attribute is not a content attribute. Per design discussion in
https://github.com/openui/open-ui/issues/1205, we decided that the best
approach was to introduce a `defaultchecked` content attribute to
control default checkedness, and then just a `checked` IDL attribute
to track and control the live state. This CL implements that behavior,
with the exception of the `defaultchecked` attribute which will come
soon.

This CL also wires up the menu item's checkability with a parent
`<fieldset>`, and managed the exclusivity checking logic inside
HTMLFieldSetElement.

This CL also wires up the `:checked` pseudo-class and `::checked`
pseudo-element for checkable menu items, and adds web platform tests
asserting the behavior and style matching of checkable menu items.

A subsequent CL will manage default styles for <legend>, and possible
horizontal rule dividers between checkable sections.

R=dizhangg

Bug: 406566432
Change-Id: I0e12ffa971e2945856937619a84578a7f1d8de9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619833
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473474}
NOKEYCHECK=True
GitOrigin-RevId: 390464001ad87e1fba54b96e2e9d39144a768b45