Store before-change style for anchored() query effects
We need to store the old style on PostStyleUpdateScope animation data in
order to trigger transitions correctly inside container-type:anchored
for the same reason we do it inside size containers. Multiple style
passes for a style change event means we need to remember the style from
the previous style change event, not rely an second-to-last update that
is stored on the Element.

Bug: 417621241
Change-Id: I469da4f5ad0fbd812d28dde58128a543775a1a05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6632801
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473654}
NOKEYCHECK=True
GitOrigin-RevId: 78870a08037cd16bc7c117b13e476195d1aabe95