Corner-shape: further tweak to color-join algorithm
Clip the tangent-bound polygon also when the corners don't intersect,
as the miter might still intersect. Adjust the fallbacks in case the miter and the edges don't intersect.

The polygon is now a hexagon, as for outlines it needs to also pass
through the start of the curve.

Also refactor a bit to use LineF and QuadraticControlPoint.

F=fmalita

Bug: 417207337
Change-Id: I339468fa04ddcab301205288a83dc634b3f79410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6634367
Commit-Queue: Noam Rosenthal <nrosenthal@google.com>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1473721}
NOKEYCHECK=True
GitOrigin-RevId: 40789269d54b23ae377769729c44327d32a41af5