[tvOS] Add support for editing text input fields
The implementation is much simpler than the iOS one because there is no
BrowserEngineKit to deal with.

Furthermore, tvOS has some peculiarities:

1. Contrary to iOS, simply adopting the UIKeyInput protocol does not
   cause the system keyboard to be displayed when RenderWidgetUIView
   becomes the first responder.
2. The system keyboard takes up the entire screen.

What we do to work around 1) is add a dummy UITextField instance that we
selectively add to the view hierarchy when we need to open or close the
on-screen keyboard.

For 2), we need to make WidgetBase::ShowVirtualKeyboardOnElementFocus()
behave like on ChromeOS and not immediately show the system keyboard
when a text element is focused. Instead, we only show it when it is
focused and a tap occurs.

Bug: 391914246, 411452047
Change-Id: Id5507b6c8b8f247b485ba76f55f330c4a2519fa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6633696
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Auto-Submit: Raphael Kubo da Costa <kubo@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1473775}
NOKEYCHECK=True
GitOrigin-RevId: 998c26e0b133bdca2c967849c39ed05112b75882