Issue DevTools events for verbose debug attribution reports
Similar to crrev.com/c/6494173, the report body is a list of objects
whose properties are not specified in the protocol itself in order to
avoid compatibility issues with respect to the Attribution Reporting
API's evolution, and because each report type has a different set of
fields.

The net/HTTP status details use the same naming scheme and format as
analogous types elsewhere in browser_protocol.pdl.

Bug: 40273989
Change-Id: I7d2356a345e5d92520e7b75fdd063960597a9a22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6634471
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474919}
NOKEYCHECK=True
GitOrigin-RevId: 74204931ebcbc68d47c3cbdfe0ad8a0e0c50b7a8