Fix actor scroll tool on smooth scrolling pages
The scroll tool would return failure if the scroll offset didn't change
but didn't consider the fact that the scroll might be animated so won't
change until the first frame of the animation.

Blink's internals already return a bool to indicate if a scroll was
started so this CL just plumbs that out to WebElement.

Bug: 411462297
Change-Id: I50f328724c831964f8f5bb699182a6d54c541c26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6647297
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475026}
NOKEYCHECK=True
GitOrigin-RevId: f0852140c5e700b94f64bdf8cdb3807c21a22a64