Delay observation after smooth scroll
The scroll tool typically scrolls instantly but if the page has smooth
scroll behavior the scroll will be animated so we don't want to take an
observation until that animation completes.

Long term it'd be more robust to have hooks into the animation system to
understand when an animation was added and when it completes. However,
this is rare enough that for now a simple delay is fine.

Fixed: 411462297
Change-Id: Ibbcae1a5b415a9a944bc4b4dcf6cd968dba572c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6647696
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475034}
NOKEYCHECK=True
GitOrigin-RevId: ad41110047ec24e112cab3bcab1b261e48fe4898