[soft navs] Avoid DOM tree walk after input/scroll
Follow-up to crrev.com/c/6641884. We stop recording LCP after input or
scroll, which now makes SoftNavigationContext::AddPaintedArea() a noop,
since AddPaintedAreaInternal() immediately returns false. But, the tree
walk which sets the context for the text/image records is not guarded by
this, which means we're doing the tree walk unnecessarily after we stop
recording LCP.

This CL changes SoftNavigationHeuristics to check if the context has
stopped recording LCP, and MaybeGetSoftNavigationContextForTiming()
returns null in that case. We could alternatively put the check in
IsNeededForTiming(), but that won't work with pre-paint-based paint
attribution (crrev.com/c/6629428).

Bug: 422754019, 330709851
Change-Id: I70139e93ff67a967c9e20c9d68b88a72625968b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6653006
Reviewed-by: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475054}
NOKEYCHECK=True
GitOrigin-RevId: 1b6b5a38f6ccc09807de137f1953ae84e6054f18