[owners] Remove weizhong@chromium.org from third_party/blink/tools/blinkpy/web_tests/web_test_analyzers/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/weizhong@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id75ac457fff0906baa07a68976b7e436d1caf3c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6652932
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475071}
NOKEYCHECK=True
GitOrigin-RevId: 1d8261c0bf41fb975348c9431933f7731acfcf99