[A11yPerformance] Correctly identify actionable nodes in AXOnScreenMode
This commit addresses an issue where some
actionable elements were not correctly identified
when the AXOnScreenMode for Switch Access was
enabled. Leaf nodes with empty bounding boxes were
being incorrectly marked as off-screen, even if
their parent containers were visible.

The logic in `MarkOnScreenNodes` is updated to
traverse up the accessibility tree from a leaf
node to find the first parent object with a
non-empty bounding box. This parent's node is then
used to determine the on-screen status, ensuring
that visually present but sizeless leaf elements
are correctly included in the accessibility tree.

A new browser test, `ActionableElementsInTree`, is
added to verify that various actionable elements
are correctly represented in the accessibility
tree when on-screen mode is active.

This change fixes the following tests when running
them with the AXModeOnScreen enabled:
- All/SwitchAccessMV2PredicateTest.IsActionable/0
- All/SwitchAccessMV2PredicateTest.IsActionable/1
- All/SwitchAccessPredicateTest.IsActionable/0
- All/SwitchAccessPredicateTest.IsActionable/1

Change-Id: Iac940376537f2b895d72439b311e4480b35c65e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6649970
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475241}
NOKEYCHECK=True
GitOrigin-RevId: 163aa29f35675d04cef21051d90e4ea640a56689