[Blink] Move test of CRH::GetRasterMode() to be testing CRCHost
I am shortly going to be moving CanvasResourceHost::GetRasterMode() to
CanvasRenderingContextHost. The one non-trivial reference to this method
on CanvasResourceHost is in a SharedGpuContext test, which creates a
FakeCanvasResourceHost to be able to test the behavior of
GetRasterMode(). This CL moves this test to be in the OffscreenCanvas
tests to ensure that the test can be preserved across the move (note:
there are no tests at the level of CanvasRenderingContextHost and it
would be slightly fiddly to add them given the number of pure virtual
methods that that class has).

Bug: 352263194
Change-Id: I8a22e0aa431a443926ebaaae0d420937c181714c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6652048
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475419}
NOKEYCHECK=True
GitOrigin-RevId: 371310c055008823cc6c308dee6797625e583098