Consider percentages in gradient default values serialization
In r1177852 it was implemented an early return to omit the default
<position> value when serializing gradient CSS Values. In that change
it's been considered just the 'center' keyword value. However, given
that 'center' computes to "50%", we should consider the
<length-percentage> to perform this check to omit the default values in
serialization.

Both radial and conic gradients have the same problem, so this CL
applies fix the issue in both cases.

Bug: 424157326
Change-Id: I515d8633d06de463a1d5db116936a3155ac7477a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6638812
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1475471}
NOKEYCHECK=True
GitOrigin-RevId: 10aa2e131dbec4719ed1cdc36938b19c98228ea7