IDB: implement web Blob storage in SQLite backing store
This change implements writing web Blobs[1] into the SQLite database as
well as reading them back out.

Since these types of blobs can be arbitrarily large they need to support
streaming, so a new helper class in `//sql` is introduced which wraps
SQLite's `sqlite3_blob`. SQLite can only handle blobs up to 2GB in size,
and we'd like IndexedDB to be able to support larger blobs. Support for
that scenario is left as a TODO.

Several new SQLite tables and triggers are added to the schema. These
additions to the schema should be self-documenting (taking into account
the adjacent comments). The detailed design[2] also contains some
discussion.

Separate helper classes are added for streamed reading and writing of
these blobs. A couple more WPT now pass. External objects that are not
of type Blob are not yet supported.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Blob

[2]
https://docs.google.com/document/d/1-N4NqdYjqaWEbfVPYnCg2Ho1u4YK48pUoQ_tSy9SUg8/edit?tab=t.0#heading=h.m7gxwxn8hw44

This rather large change increases Fuchsia binary size a bit. We'll
eventually be able to delete some of the LevelDB backing store code
it replaces, so this is temporary, but also just due to the large
number of LOC and not due to pulling in new deps.

Fuchsia-Binary-Size: See above

Bug: 419208485

Change-Id: Ia73ab7d65d81486db1a21b02d5ba78f5466af90c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6613582
Reviewed-by: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Commit-Queue: Evan Stade <evanstade@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1475631}
NOKEYCHECK=True
GitOrigin-RevId: 5032c2708ae5e3027c6ea9f1e48f0fae4b46b0e1