[a11y] Prevent infinite loop in `ComputeNodesOnLine`
This fix uses the maximum number of runs
(`kMaxInlineCursorNextObjectCalls`) to the
`do-while` loop, breaking the loop if the maximum
number of attempts to navigate to the next object
or line is reached. A further investigation is
being done to identify the root cause of the issue
on the `InlineCursor` side, which is the expected
long term fix.

R=aleventhal@chromium.org

Bug: 417951578
Change-Id: I3a13b94a9018131f156357d4008977c38b5b4f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651216
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1475727}
NOKEYCHECK=True
GitOrigin-RevId: 88ecd2fa8d58c5a5c3562ce324003cf915077b1d