[Editing] Traverse flat tree in delete selection to handle slot distribution
|MergeParagraphs| previously used DOM tree traversal to identify the
paragraph boundaries for moving content. This approach lead to an
incorrect end position in <details> that utilize shadow DOM and
slots. The children of it, distributed via slots, were not
correctly recognized as part of shadow tree.
This patch addresses the problem by shifting to flat tree traversal
to find the start and end positions.

Bug: 413595949
Change-Id: Ie21164ad1be902b1c7bc4c6430b246544bc5e9f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6620033
Commit-Queue: Shweta Bindal <shwetabindal@microsoft.com>
Reviewed-by: Sambamurthy Bandaru <sambamurthy.bandaru@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476000}
NOKEYCHECK=True
GitOrigin-RevId: ac422a1657fe7e817e8ed4967366e392ac484725