[Blink] Introduce and use CanvasRenderingContext::IsAccelerated()
Use this method to abstract the check on the resource provider being
accelerated in CanvasRenderingContextHost::IsAccelerated(). This will
enable changing the implementation of this check incrementally in
individual contexts in followup work to eliminate dependencies
on CanvasResourceProvider.

In the medium term we might replace this whole flow with checking
whether the image is texture-backed at [1][2] under a killswitch, but
I decided to separate that stream of work from the one that we're
engaged on here to avoid blocking the current workstream.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/canvas/canvas2d/canvas_2d_recorder_context.cc;l=2776;drc=4388d13dc2a92e9123085050cf95db052cba8ab3
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/canvas/canvas2d/canvas_2d_recorder_context.cc;l=2465;drc=4388d13dc2a92e9123085050cf95db052cba8ab3

Bug: 352263194
Change-Id: I9e6061fdd73abe67e87ba60fb834c506f8561c61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651559
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476030}
NOKEYCHECK=True
GitOrigin-RevId: 2ff1ae882807de4ed4b2f0ff11bd9d5f4e65aa9a