Add IsParserResumingCalledBeforePausing UMA
We observed the yielded parser is resumed by timeout, not by the user
timing API call. We initially thought this is rare case, but it turned
out that it's much more frequent than we expected from the early
analysis.

Our hypothesis is that the order of the user timing API calls is the
reverse order of the expected one. We assume in some cases the parser
resume signal is sent before the parser pause signal. If so, the first
parser resume call is no-op, then the parser is paused after the target
milesone, then the parser is resumed by the timeout.

This CL idnetifies such case by adding a boolean histogram.

Bug: 425962649
Change-Id: Ic8fa89281f08e2f86d589f5ec18552e30ec66fee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6656240
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Jiacheng Guo <gjc@google.com>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476034}
NOKEYCHECK=True
GitOrigin-RevId: 8af14c9bdc88f1a0975c23e63c0834d75b06c770