Move TextDecorationInfo::Bounds() to DecorationLinePainter
This is where we want to end up eventually, and moving it will allow to
reuse some of the helpers from the latter.

Change the new function to use StrokeStyle values instead of
ETextDecorationStyle. Fold BoundsForWavy() and make
BoundsForDottedOrDashed() public for the time being (to be removed in a
follow-up CL).

Bug: 425866014
Change-Id: I0ec092b5efcb25a6fbe0fb6d745c77beb478089a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6653659
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1476050}
NOKEYCHECK=True
GitOrigin-RevId: f3a0cfbf652f9979f9742ecd297a1353209e8320