Query anchor(fallback) values instead of indices
Per explainer[1] and CSSWG discussion[2]:

The extraction of the position-try declaration block and try-tactics is
moved into UpdateStyleAndLayoutTreeForOutOfFlow() since the whole
PositionTryFallback now needs to be available for updating the
CSSContainerValues for querying anchored(fallback).

Writing-direction tests will be added in another CL. The tests in this
CL are just changes for moving from index-based to value-based matching.

[1] https://github.com/w3c/csswg-drafts/blob/main/css-anchor-position-1/anchored_container_query.md
[2] https://github.com/w3c/csswg-drafts/issues/8171#issuecomment-2963456372

Bug: 417621241

Change-Id: I98a318cc41e0f26f123e67e26e3b24387c4da179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6650679
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476070}
NOKEYCHECK=True
GitOrigin-RevId: 63511a4f6879583cd1b3143fd4e25c1516033031