Clamp the result of the CSS progress() function
This CL addresses https://github.com/w3c/csswg-drafts/issues/11825
by clamping the result of the function by default, but doesn't add
the unclamped keyword as it's being bickeshed now.

Also, test changes to container-progress() and media-progress()
tests are on hold before the bikesheding of the keyword is done,
those functions are to be dropped from CSS anyway.

The purpose of this CL is to update the default behavior asap,
as progress() is already shipped.

Bug: 425929508
Change-Id: I0145787fba6d70d4c3c286e5afaaef4c6d6ff762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6654966
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476125}
NOKEYCHECK=True
GitOrigin-RevId: 7ad6aea7dbefa19264b5ffa449d1c0d30298c4e0