[blink scheduler] Clean up kThrottleTimedOutIdleTasks feature
This CL removes the kThrottleTimedOutIdleTasks kill switch, which was
added in M134. It also removes PrioritisationType::kJavaScriptTimer,
which is no longer used (it used to affect queue policy, but that was
removed in crrev.com/c/3791718). I'm including this in this patch to
simplify setting up the throttleable types.

Bug: 40180881
Change-Id: Ifdc8531f215b5f0c552d15306678aaec5bc2667a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651044
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476157}
NOKEYCHECK=True
GitOrigin-RevId: 3bb50fff0b0b9a943d1a3373362fe2c8fe37521d