[ios] Fix fallout of tests building webnfc
Tests used a different toolchain that the chrome or content_shell
builds which was built with different flags. Bring the test code in
line with the build of the rest of the code so it doesn't have to
be built multiple times with different toolchains.

Bug: 40254930
Change-Id: I186c65d61097e48b4d0c511f07d663e696bec15d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6652622
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476216}
NOKEYCHECK=True
GitOrigin-RevId: 5e329778ed59fc9310027e8bee9388ad9d52bd14