[soft navs] Add soft LCP recording to UKM.
This patch moves the LargestContentfulPaintDetails for Soft Navs from
the global Paint Timing Detector into the Soft Navigation Context, where
the Interaction specific LCP calculator now lives. Now that we never
update FP/FCP/LCP in the global paint timing detector for soft-navs, we
can continue to remove references to soft-navs resetting there.

The SoftNavigationHeuristics class decides which context is the "current
context" for UKM recording purposes, and that is based on the current
URL. Other context's might be tracking paints but will not be updating
or reporting LCP candidates to UKM. Not until we get Pre-Paint tree
walking, and migrate from soft-LCP entry to some
InteractionContentfulPaint entry, and adjust UkmPLMO to support such
reporting.

As part of this change, we now check if we actually have a new largest
image or largest text candidate, and only bother to update performance
timeline or ukm if one of these changed. Previously we would would
update every new attributable paint, which would end up no-op further
down in LCP reporting.

We no longer delete the LCP calculator inside each Context after Input
or Scroll. This is because the specific LCP values, and metrics
reporting, might continue after this time. We might be able to delete
the lcp calculator after the last image/text candidate have been
measured and reported, since we copy LCP metrics data, but we already
expect to experiment with paint reporting after input or scroll (even if
this doesn't emit LCP candidates) in order to support scroll restoration
and/or "Interaction to Visually Complete" use cases (potentially).

Bug: 425398556, 382245480
Change-Id: Ie3030ddc2f14c59912984d912fbe0de1f22ae5c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6648249
Commit-Queue: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476231}
NOKEYCHECK=True
GitOrigin-RevId: 462d66aafcee5855305083fe75a1469d8b29cb1f